+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/ContinueArguments.java
@@ -21,7 +21,7 @@ public class ContinueArguments {
    * but will continue on all threads, it should set the allThreadsContinued attribute in the response to true.
    */
   @NonNull
-  private Integer threadId;
+  private Long threadId;
   
   /**
    * Continue execution for the specified thread (if possible). If the backend cannot continue on a single thread
@@ -29,7 +29,7 @@ public class ContinueArguments {
    */
   @Pure
   @NonNull
-  public Integer getThreadId() {
+  public Long getThreadId() {
     return this.threadId;
   }
   
@@ -37,7 +37,7 @@ public class ContinueArguments {
    * Continue execution for the specified thread (if possible). If the backend cannot continue on a single thread
    * but will continue on all threads, it should set the allThreadsContinued attribute in the response to true.
    */
-  public void setThreadId(@NonNull final Integer threadId) {
+  public void setThreadId(@NonNull final Long threadId) {
     this.threadId = threadId;
   }
