+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/OutputEventArguments.java
@@ -40,7 +40,7 @@ public class OutputEventArguments {
    * <p>
    * This is an optional property.
    */
-  private Integer variablesReference;
+  private Long variablesReference;
   
   /**
    * An optional source location where the output was produced.
@@ -54,14 +54,14 @@ public class OutputEventArguments {
    * <p>
    * This is an optional property.
    */
-  private Integer line;
+  private Long line;
   
   /**
    * An optional source location column where the output was produced.
    * <p>
    * This is an optional property.
    */
-  private Integer column;
+  private Long column;
   
   /**
    * Optional data to report. For the 'telemetry' category the data will be sent to telemetry, for the other
@@ -117,7 +117,7 @@ public class OutputEventArguments {
    * This is an optional property.
    */
   @Pure
-  public Integer getVariablesReference() {
+  public Long getVariablesReference() {
     return this.variablesReference;
   }
   
@@ -127,7 +127,7 @@ public class OutputEventArguments {
    * <p>
    * This is an optional property.
    */
-  public void setVariablesReference(final Integer variablesReference) {
+  public void setVariablesReference(final Long variablesReference) {
     this.variablesReference = variablesReference;
   }
   
@@ -156,7 +156,7 @@ public class OutputEventArguments {
    * This is an optional property.
    */
   @Pure
-  public Integer getLine() {
+  public Long getLine() {
     return this.line;
   }
   
@@ -165,7 +165,7 @@ public class OutputEventArguments {
    * <p>
    * This is an optional property.
    */
-  public void setLine(final Integer line) {
+  public void setLine(final Long line) {
     this.line = line;
   }
   
@@ -175,7 +175,7 @@ public class OutputEventArguments {
    * This is an optional property.
    */
   @Pure
-  public Integer getColumn() {
+  public Long getColumn() {
     return this.column;
   }
   
@@ -184,7 +184,7 @@ public class OutputEventArguments {
    * <p>
    * This is an optional property.
    */
-  public void setColumn(final Integer column) {
+  public void setColumn(final Long column) {
     this.column = column;
   }
